Factor out ServeFD & systemd in server for Linux only (Not Windows)
