+++ api/server/server_windows.go
@@ -18,3 +18,14 @@ func NewServer(proto, addr string, job *engine.Job) (Server, error) {
 		return nil, errors.New("Invalid protocol format. Windows only supports tcp.")
 	}
 }
+
+// Called through eng.Job("acceptconnections")
+func AcceptConnections(job *engine.Job) engine.Status {
+
+	// close the lock so the listeners start accepting connections
+	if activationLock != nil {
+		close(activationLock)
+	}
+
+	return engine.StatusOK
+}
